v2.0-baseline/requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Deep learning
torch==2.0.1
transformers==4.30.2
sentence-transformers==2.2.2

# Vector search
faiss-cpu==1.7.4

# Utilities
pyyaml==6.0
python-dotenv==1.0.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Data validation
pydantic==2.1.1

# API clients
openai==0.27.8
anthropic==0.3.0
